Reorganize tests and streamline editor structure
