+++ tests/vad-test.ts
